adjusted search_bar hidden_field selection for nil option related to Restaurant#index search

+++ app/models/grocery/importer.rb
@@ -9,7 +9,7 @@ class Grocery::Importer
     @erecipe_id = grocery_params.fetch(:erecipe_id)
 
     @name = grocery_params.fetch(:name)
-    @quantity = grocery_params.fetch(:quantity).to_i
+    @quantity = grocery_params.fetch(:quantity).to_f
     @measurement = grocery_params.fetch(:measurement)
     @category = grocery_params.fetch(:category)
     @list_add = grocery_params.fetch(:list_add).to_i
